requirements.txt
streamlit
torch
timm
torchvision
Pillow
